Hito 8: dos bugs reales encontrados en la primera prueba end-to-end contra un servidor real
1. TypeError en cualquier entorno con .env real (config/conversations.php)

   env('CONVERSATION_MESSAGE_DEDUP_HOURS', 48) devuelve el default de PHP
   tal cual (un int) solo cuando la variable de entorno NO existe. En
   cuanto SÍ está seteada — como en cualquier despliegue real, a
   diferencia de este entorno de test donde nunca se había seteado — env()
   siempre devuelve un string, y Carbon::addHours()/addMinutes() exigen
   int|float. El pipeline entero fallaba con TypeError apenas llegaba un
   mensaje real vía el webhook. Cast (int) explícito en ambos valores
   (message_dedup_hours, continuity_ttl_minutes) — mismo problema en los
   dos. Regresión: tests/Feature/ConversationsConfigTest.php, reproduce la
   condición real (env seteada como string) requiriendo el archivo de
   config directamente, no vía config() ya cacheado por Laravel.

2. WeeklyScheduleFieldExtractor guardaba weekday en convención equivocada

   Le pedía a la IA 1=lunes..7=domingo (ISO), pero resource_schedules.weekday
   (migración Hito 1) y AvailabilityCalculator (Carbon::dayOfWeek) usan
   0=domingo..6=sábado. Lunes a sábado coinciden numéricamente en ambas
   convenciones por pura casualidad — domingo es el único día que diverge
   (7 guardado vs 0 consultado) — así que quedó invisible en todos los
   tests anteriores hasta que "mañana" cayó en domingo durante esta
   sesión. Impacto real: cualquier negocio registrado por WhatsApp
   quedaba con el domingo silenciosamente sin disponibilidad, sin importar
   qué horario hubiera configurado. Corregido el prompt de extracción;
   agregado un test determinista (fecha explícita al próximo domingo, no
   "mañana") para no volver a depender de qué día corra la suite.

Verificado: suite completa 194/194 tests en verde (434 aserciones), Pint
limpio. Ambos bugs se corrigieron y probaron localmente antes de
redesplegar a staging.

+++ config/conversations.php
@@ -15,7 +15,13 @@ return [
     |
     */
 
-    'message_dedup_hours' => env('CONVERSATION_MESSAGE_DEDUP_HOURS', 48),
+    // (int) explícito: env() siempre devuelve string cuando la variable SÍ
+    // está seteada en un .env real (solo cae al default tal cual — un int
+    // de PHP — cuando la variable no existe). Sin el cast, Carbon::addHours()
+    // revienta con TypeError en cualquier entorno con .env real (encontrado
+    // en el Hito 8, invisible en tests porque nunca se seteó esta variable
+    // en un .env de verdad hasta ese momento).
+    'message_dedup_hours' => (int) env('CONVERSATION_MESSAGE_DEDUP_HOURS', 48),
 
     /*
     |--------------------------------------------------------------------------
@@ -33,6 +39,7 @@ return [
     |
     */
 
-    'continuity_ttl_minutes' => env('CONVERSATION_CONTINUITY_TTL_MINUTES', 240),
+    // Mismo motivo del cast (int) que message_dedup_hours arriba.
+    'continuity_ttl_minutes' => (int) env('CONVERSATION_CONTINUITY_TTL_MINUTES', 240),
 
 ];
